+++ docs/news.rst
@@ -7,3 +7,4 @@ Changes
 * Initial release.
 * Support for getting data usage.
 * Support for getting current balance.
+* Support for sending messages.